+++ example.js
@@ -4,18 +4,18 @@ mongolass.connect('mongodb://localhost:27017/test')
 
 const cache = require('.')({
   prefix: 'cache:',
-  expire: 10 * 1000 // default expire
+  expire: 10 * 60 * 1000 // default expire
 })
 
-const User = mongolass.model('User')
+const User = mongolass.model('UserTest')
 
 ;(async function () {
-  // create some users
+  // create test users
   for (let i = 1; i <= 10; i++) {
     await User.insertOne({ name: i })
   }
 
-  const getUsersByPage = cache(function getUsersByPage(p) {
+  const getUsersByPage = cache(function getUsersByPage (p) {
     return User
       .find()
       .skip((p - 1) * 1)
@@ -32,8 +32,13 @@ const User = mongolass.model('User')
   await getUsersByPage(1)
   await getUsersByPage(2)
   await getUsersByPage(2)
-  console.log(await getUsersByPage(3))
+  await getUsersByPage(3)
 
+  await getUsersByPage.clear(1) // clear cache
+  await getUsersByPage.clear(2) // clear cache
+  await getUsersByPage.clear(3) // no effect, because there is no cache
+
+  // remove test users
   await User.remove()
 
   process.exit()